Move process_docstring to separate file
