+++ conf.in.py
@@ -260,10 +260,10 @@ def linkcode_resolve(domain, info):
 def setup(app):
     try:
         from autoautosummary import AutoAutoSummary
+        from docs_builder.autodoc import AutoDocAdditions
         from docs_builder.documenters import OverloadedPythonMethodDocumenter
         from process_links import (
             process_bases,
-            process_docstring,
             process_signature,
             skip_member,
         )
@@ -271,7 +271,7 @@ def setup(app):
         app.add_directive("autoautosummary", AutoAutoSummary)
         app.add_autodocumenter(OverloadedPythonMethodDocumenter)
         app.connect("autodoc-process-signature", process_signature)
-        app.connect("autodoc-process-docstring", process_docstring)
+        app.connect("autodoc-process-docstring", AutoDocAdditions.process_docstring)
         app.connect("autodoc-skip-member", skip_member)
         app.connect("autodoc-process-bases", process_bases)
     except BaseException as e:
